refactor: migrate to Vite and update project structure
- Migrate from CJS to ESM configuration (vite.config.cjs -> vite.config.js)
- Reorganize assets structure and remove unused media files
- Update package dependencies and configuration files
- Move index.html to root directory
- Update Navbar component and Constants
- Clean up project structure and remove redundant files

This commit modernizes the project setup and improves organization
