+++ postcss.config.js
@@ -1,4 +1,6 @@
-export const plugins = {
-  '@tailwindcss/postcss': {},
-  autoprefixer: {},
-};
+export default {
+  plugins: {
+    tailwindcss: {},
+    autoprefixer: {},
+  },
+}
