+++ src/main.jsx
@@ -4,7 +4,7 @@ import App from './App'; // App.jsx must exist in `src`
 import './index.css';    // index.css must exist in `src`
 
 createRoot(document.getElementById('root')).render(
-  <StrictMode>c
+  <StrictMode>
 	<App />
   </StrictMode>
 );
