--- vite.config.cjs
@@ -1,7 +1,0 @@
-import { defineConfig } from 'vite';
-import react from '@vitejs/plugin-react';
-
-export default defineConfig({
-  plugins: [react()],
-  root: './src', // Adjust if necessary
-});
